Added shuffle folder, displayNamePrefix
displayNamePrefix is a property in DLNAResource that allows you to add
a string to the beginning of the calculated display name. This is used
by ShuffleFolder, which adds a number to the beginning of the display
name to prevent DLNA clients from sorting the list again.

ShuffleFolder is a virtual folder that shuffles it's children each time
it's resolved.