first attempt to include a internal notification center
Instead of e.g. having to update the status bar message from within the
core components with PMS.get().getFrame().setStatusLine(statusMsg) it
would be nice if a message could be posted to a message queue and
subscribers would get notified and act accordingly. Same goes for
Player which could load the plugins when it's being notified they have
been loaded