NotificationCenter v2
Made the notification center generic. Specify the expected notification
object type when getting a singleton instance
(NotificationCenter.getInstance(Object.class)). The class specifies the
queue to subscribe to as well as the received object when a message is
being posted.
Replaced the plugging calling for StartStopListeners from DlnaResource
to StartStopNotifier who listens to the new notifications.
Load plugins in GUI either on start up if available or by notification.
Load FinalizeTranscoderArgsListener upon notification in Player
Remove plugins from general tab.